GEODE-9564: Rename package from ParameterRequirements to parameters
Authored-by: Donal Evans <doevans@vmware.com>
