+++ geode-apis-compatible-with-redis/src/main/java/org/apache/geode/redis/internal/RedisCommandType.java
@@ -20,16 +20,6 @@ import static org.apache.geode.redis.internal.RedisCommandSupportLevel.SUPPORTED
 import static org.apache.geode.redis.internal.RedisCommandSupportLevel.UNSUPPORTED;
 import static org.apache.geode.redis.internal.RedisConstants.ERROR_SYNTAX;
 
-import org.apache.geode.redis.internal.ParameterRequirements.ClusterParameterRequirements;
-import org.apache.geode.redis.internal.ParameterRequirements.EvenParameterRequirements;
-import org.apache.geode.redis.internal.ParameterRequirements.ExactParameterRequirements;
-import org.apache.geode.redis.internal.ParameterRequirements.MaximumParameterRequirements;
-import org.apache.geode.redis.internal.ParameterRequirements.MinimumParameterRequirements;
-import org.apache.geode.redis.internal.ParameterRequirements.OddParameterRequirements;
-import org.apache.geode.redis.internal.ParameterRequirements.ParameterRequirements;
-import org.apache.geode.redis.internal.ParameterRequirements.SlowlogParameterRequirements;
-import org.apache.geode.redis.internal.ParameterRequirements.SpopParameterRequirements;
-import org.apache.geode.redis.internal.ParameterRequirements.UnspecifiedParameterRequirements;
 import org.apache.geode.redis.internal.executor.Executor;
 import org.apache.geode.redis.internal.executor.RedisResponse;
 import org.apache.geode.redis.internal.executor.UnknownExecutor;
@@ -139,6 +129,16 @@ import org.apache.geode.redis.internal.executor.string.SetRangeExecutor;
 import org.apache.geode.redis.internal.executor.string.StrlenExecutor;
 import org.apache.geode.redis.internal.netty.Command;
 import org.apache.geode.redis.internal.netty.ExecutionHandlerContext;
+import org.apache.geode.redis.internal.parameters.ClusterParameterRequirements;
+import org.apache.geode.redis.internal.parameters.EvenParameterRequirements;
+import org.apache.geode.redis.internal.parameters.ExactParameterRequirements;
+import org.apache.geode.redis.internal.parameters.MaximumParameterRequirements;
+import org.apache.geode.redis.internal.parameters.MinimumParameterRequirements;
+import org.apache.geode.redis.internal.parameters.OddParameterRequirements;
+import org.apache.geode.redis.internal.parameters.ParameterRequirements;
+import org.apache.geode.redis.internal.parameters.SlowlogParameterRequirements;
+import org.apache.geode.redis.internal.parameters.SpopParameterRequirements;
+import org.apache.geode.redis.internal.parameters.UnspecifiedParameterRequirements;
 
 /**
  * The redis command type used by the server. Each command is directly from the redis protocol.
