+++ geode-apis-compatible-with-redis/src/main/java/org/apache/geode/redis/internal/parameters/SlowlogParameterRequirements.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package org.apache.geode.redis.internal.ParameterRequirements;
+package org.apache.geode.redis.internal.parameters;
 
 import static org.apache.geode.redis.internal.RedisConstants.ERROR_NOT_INTEGER;
 import static org.apache.geode.redis.internal.RedisConstants.ERROR_UNKNOWN_SLOWLOG_SUBCOMMAND;
